Moved CWorldSearch class into its own cpp/h files.
