+++ src/game/clients/CClientEvent.cpp
@@ -11,8 +11,9 @@
 #include "../CSector.h"
 #include "../CServer.h"
 #include "../CWorld.h"
-#include "../CWorldMap.h"
 #include "../CWorldGameTime.h"
+#include "../CWorldMap.h"
+#include "../CWorldSearch.h"
 #include "../spheresvr.h"
 #include "../triggers.h"
 #include "CClient.h"
